Alot of new things.
Redesigned the GUI so that the events log is within the main window for easier visuals. Added threading to the startNewSequence function so that both the variables in the main GUI and calculations are performed simultaneously. Took out the save excel command and added it back as a user function (i.e. user presses "Save Data: button to save any data).