Safety Proofed the Code + other features.
Made sure the code was safety proof. Such as make sure the user cannot run tests if the instruments havnt been setup etc. Added a query validation feature which can be chosen in the GUI. Changed the TextToVoice feature so that the user can select if user wants to have this feature left on. Note that TextToVoice add approximately 22sec to each reading.